re #103 move epics client impl details back to cpp file
